refactor: 승자를 찾는 로직 RacingGame객체로 이동
