+++ src/main/java/racingcar/model/Cars.java
@@ -9,8 +9,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Cars {
-    public static final int MINIMUM_POSITION = 0;
-    public static final String JOIN_BY_COMMA = ", ";
     private static final String COMMA = ",";
     private static final String BLANK = "";
     private static final String SPACE = " ";
@@ -46,25 +44,6 @@ public class Cars {
         cars.forEach(car -> {
             car.move(RandomNumberGenerator.generate());
         });
-    }
-
-    public String findWinner() {
-        int maxPosition = findMaxPosition();
-        return findWinnerName(maxPosition);
-    }
-
-    private int findMaxPosition() {
-        return cars.stream()
-                .map(Car::getPosition)
-                .max(Comparator.comparing(x -> x))
-                .orElse(MINIMUM_POSITION);
-    }
-
-    private String findWinnerName(int maxPosition) {
-        return cars.stream()
-                .filter(car -> maxPosition == car.getPosition())
-                .map(Car::getName)
-                .collect(Collectors.joining(JOIN_BY_COMMA));
     }
 
     public List<Car> getCars() {
