+++ src/test/java/racingcar/CarsTest.java
@@ -55,15 +55,4 @@ public class CarsTest {
 
         assertThat(expected).isEqualTo(actual);
     }
-
-    @Test
-    void 최종_우승자_출력() {
-        List<Car> carList = new ArrayList<>();
-        carList.add(new Car("bom", 3));
-        carList.add(new Car("sun", 4));
-        Cars cars = new Cars(carList);
-
-        String winnerName = cars.findWinner();
-        assertThat(winnerName).contains("sun").doesNotContain("bom");
-    }
 }
